refactor: cam into separate file
