+++ pocovidnet/requirements.txt
@@ -2,5 +2,6 @@ tensorflow==2.2.0rc1
 scikit-learn
 matplotlib
 imutils
+numpy>=1.18.2
 opencv-contrib-python
 flask
